+++ src/Attestation/Attestation.Test/ScenarioTests/AttstationPolicySignerTests.cs
@@ -20,57 +20,36 @@ using Xunit;
 
 namespace Microsoft.Azure.Commands.Attestation.Test.ScenarioTests
 {
-    public class AttstationTests : RMTestBase
+    public class AttestationPolicySignerTests : RMTestBase
     {
         public XunitTracingInterceptor _logger;
 
-        public AttstationTests(Xunit.Abstractions.ITestOutputHelper output)
+        public AttestationPolicySignerTests(Xunit.Abstractions.ITestOutputHelper output)
         {
             _logger = new XunitTracingInterceptor(output);
             XunitTracingInterceptor.AddToContext(_logger);
             TestExecutionHelpers.SetUpSessionAndProfile();
         }
 
-        #region New-AzureRmAttestation        
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
-        public void TestCreateAttestation()
+        public void TestGetAttestationPolicySigners()
         {
-            AttestationController.NewInstance.RunPowerShellTest(_logger, "Test-CreateAttestation");
+            AttestationController.NewInstance.RunDataPowerShellTest(_logger, "Test-GetAttestationPolicySigners");
         }
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
-        public void TestCreateAttestationWithPolicySigningCertificate()
+        public void TestAddAttestationPolicySigner()
         {
-            AttestationController.NewInstance.RunPowerShellTest(_logger, "Test-CreateAttestationWithPolicySigningCertificate");
+            AttestationController.NewInstance.RunDataPowerShellTest(_logger, "Test-AddAttestationPolicySigner");
         }
-
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
-        public void TestCreateAttestationWithTags()
+        public void TestRemoveAttestationPolicySigner()
         {
-            AttestationController.NewInstance.RunPowerShellTest(_logger, "Test-CreateAttestationWithTags");
+            AttestationController.NewInstance.RunDataPowerShellTest(_logger, "Test-RemoveAttestationPolicySigner");
         }
-        #endregion
-
-        #region Get-AzureRmAttestation
-        [Fact]
-        [Trait(Category.AcceptanceType, Category.CheckIn)]
-        public void TestGetAttestation()
-        {
-            AttestationController.NewInstance.RunPowerShellTest(_logger, "Test-GetAttestation");
-        }
-        #endregion
-
-        #region Remove-AzureRmAttestation
-        [Fact]
-        [Trait(Category.AcceptanceType, Category.CheckIn)]
-        public void TestDeleteAttestationByName()
-        {
-            AttestationController.NewInstance.RunPowerShellTest(_logger, "Test-DeleteAttestationByName");
-        }
-        #endregion
     }
 }