+++ src/ApiManagement/ApiManagement.ServiceManagement/Commands/NewAzureApiManagementIdentityProvider.cs
@@ -88,6 +88,12 @@ namespace Microsoft.Azure.Commands.ApiManagement.ServiceManagement.Commands
             HelpMessage = "Password Reset Policy Name. Only applies to AAD B2C Identity Provider. This parameter is optional.")]
         public String PasswordResetPolicyName { get; set; }
 
+        [Parameter(
+            ValueFromPipelineByPropertyName = true,
+            Mandatory = false,
+            HelpMessage = "Signin Tenant to use for AAD Authentication. Only applies to AAD B2C Identity Provider. This parameter is optional.")]
+        public String SigninTenant { get; set; }
+
         public override void ExecuteApiManagementCmdlet()
         {
             if (ShouldProcess(Type.ToString("g"), "Creates a new Identity Provider"))
@@ -102,7 +108,8 @@ namespace Microsoft.Azure.Commands.ApiManagement.ServiceManagement.Commands
                     SigninPolicyName,
                     SignupPolicyName,
                     PasswordResetPolicyName,
-                    ProfileEditingPolicyName);
+                    ProfileEditingPolicyName,
+                    SigninTenant);
 
                 WriteObject(identityProvider);
             }